requirements.txt
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
phonenumberslite==8.13.24
psycopg[c]==3.1.16
requests==2.31.0
requests-ntlm==1.2.0
scikit-learn==1.3.2
uwsgi==2.0.23
whitenoise==6.6.0
xlsxwriter==3.1.9

# Django
django==4.2.7
django-cors-headers==4.3.0
django-environ==0.11.2
django-grappelli==3.0.8
django-import-export==3.3.1
django-leaflet==0.29.0
django-phonenumber-field==7.2.0
djangorestframework==3.14.0